Move GameView to standalone file
